Refactor iterators_test for fewer dependencies
In one place in the iterators_test, I use the All function, which
depends on the mapped iterator, as well as uses an untested function. To
reduce complication, I thought it better not to depend on those things
in the test so that if either of those are broken, the test is still
valid.