Handle require on complete dependencies
If a task requires another task which has already completed, then we
will no longer add the dependency to the requirer's dependency list so
that the dependent task does not wait on an already completed task.

Also just changed the example's number of parallel tasks to 1 just to
simplify debugging for now. 2 makes sense in the future once things
actually work.