Rework communication into one main channel
Problem: If communication of different message types are performed
across several channels, then messages may be received out of order.
This will break functionality such as waiting if for example a request
to wait is received before the declared dependencies for the task.

Solution: Messages which need to be serialized in a specific order are
now sent on the same channel. Messages that can be sent asynchronously
use a different channel.