Refactor into roles
There will be four roles going forward:

manager: Single instance, performs scheduling, source of truth for task
resolution. Long-lived.
supervisor: Short-lived, acts as a communication relay between the
manager and a worker, communicating with the worker via the worker's
handler.
handler: Relays information between the supervisor and the worker. One
handler exists for each worker, whether the worker is currently
executing or not.
worker: Unit that executes the task.