Add code generation idea
Had an idea for reducing the amount of boilerplate that users would need
to write to use nbt, as well as make it look nicer.

The idea would be to use a format similar to Rake (but in Go, and with
our dynamic require statements), and then simply generate the required
task types to make it work.