More generator progress
The generator package has been worked on some more, to the point where
it should now be able to get tested, and I've tried running it.
Unfortunately, I've hit a snag trying to get the package name from an
import path, so I'm looking into ways that this can get done.