Add helper file generation
The helper file that declares constants necessary for the other files is
now generated, and things are almost to 100% functional.

What's currently missing is a safe way to import names into files
without possibly conflicting with existing imports that the user
included. This will be a problem to look for with the `nbt` name, as
well as `binary` and `fnv`.

One potential solution to these issues would be to define as much as
possible in the helper file, whose imports we control entirely, and then
use those helper functions. This should work for everything but nbt, but
dealing with just nbt is a lot easier than dealing with them all.

In fact, we could declare these helpers under nbt and then not even
generate them.