Add generation of named tasks
Tasks which can be named on the command line are now automatically
detected, and a function to create a map[string]ntr.TaskSupplier for
them will now be generated.

The plan is for this function to be used either in a generated main
function, or in one that the user writes. In the latter case, the user
may use the function for convenience, or do their own thing.