Add ntrmain
ntrmain provides a common main function to be used when one isn't
provided by the user.

Currently, generating it does work, although disabling it is not yet
implemented fully. It should be the case that when the user has their
own main function, the generated main function should not be made, but
currently we do not propagate a taskFunc instance for the main function
so this will never happen.